+++ tests/Engine/Tool/PageSpeedInsights/RulesetTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace PageExperience\Tests\Engine\Tool\Lighthouse;
+namespace PageExperience\Tests\Engine\Tool\PageSpeedInsights;
 
-use PageExperience\Engine\Tool\Lighthouse\Ruleset;
+use PageExperience\Engine\Tool\PageSpeedInsights\Ruleset;
 use PageExperience\Tests\TestCase;
 
 /**
- * Test the Lighthouse Ruleset class.
+ * Test the PageSpeedInsights Ruleset class.
  *
  * @package ampproject/px-toolbox-php
  */
